+++ pkg/deploy/apply.go
@@ -51,6 +51,7 @@ var (
 	deleteBeforeApplyAnnotation = resourceutil.GetMiaAnnotation("delete-before-apply")
 	awaitCompletionAnnotation   = resourceutil.GetMiaAnnotation("await-completion")
 	apply                       = decorateApplyResource(withAwaitableResource, withDeletableResource)
+	getTimeNowSeconds           = func() time.Time { return time.Now().Round(time.Second) }
 )
 
 func decorateApplyResource(decorators ...func(applyFunction) applyFunction) applyFunction {
@@ -98,7 +99,7 @@ func withAwaitableResource(apply applyFunction) applyFunction {
 		// register a watcher and starts to listen for events for the gvr
 		// if res is annotated with awaitCompletionAnnotation
 		var watchEvents <-chan watch.Event
-		startTime := time.Now()
+		startTime := getTimeNowSeconds()
 		awaitCompletionValue, awaitCompletionFound := res.Object.GetAnnotations()[awaitCompletionAnnotation]
 		if awaitCompletionFound {
 			watcher, err := clients.dynamic.Resource(gvr).
@@ -136,7 +137,10 @@ func withAwaitableResource(apply applyFunction) applyFunction {
 		// consume watcher events and wait for the resource to complete or exit because of timeout
 		for {
 			select {
-			case event := <-watchEvents:
+			case event, ok := <-watchEvents:
+				if !ok {
+					return errors.New("Watch channel closed unexpectedly")
+				}
 				isCompleted, err := handleResourceCompletionEvent(res, &event, startTime)
 				if err != nil {
 					msg := "Error while watching resource events"
@@ -190,7 +194,7 @@ func handleResourceCompletionEvent(res resourceutil.Resource, event *watch.Event
 			return false, nil
 		}
 		// check if job has completed after start time
-		if completedAt := jobFromEvent.Status.CompletionTime; completedAt != nil && completedAt.Time.After(startTime) {
+		if completedAt := jobFromEvent.Status.CompletionTime; completedAt != nil && !completedAt.Time.Before(startTime) {
 			fmt.Println("Job completed:", jobFromEvent.Name)
 			return true, nil
 		}
@@ -219,7 +223,7 @@ func handleResourceCompletionEvent(res resourceutil.Resource, event *watch.Event
 			return false, nil
 		}
 
-		if refreshedAt := extsecFromEvent.Status.RefreshTime; refreshedAt.Time.After(startTime) {
+		if refreshedAt := extsecFromEvent.Status.RefreshTime; !refreshedAt.Time.Before(startTime) {
 			fmt.Println("ExternalSecret completed:", extsecFromEvent.Name)
 			return true, nil
 		}