+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- support for jobs annotation `mia-platform.eu/await-completion` for waiting job completion after it has been applied on the cluster
+
 ## [1.1.0] - 2022-03-17
 
 ### Changed