+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - default order for `SecretProviderClass` resource kind
 - new annotation `mia-platform.eu/apply-before-kinds` to override default resources application order
 - support for jobs annotation `mia-platform.eu/await-completion` for waiting job completion after it has been applied on the cluster
-- support for  resource`ExternalSecrets` for the annotation `mia-platform.eu/await-completion`
+- support for resource `ExternalSecrets` for the annotation `mia-platform.eu/await-completion`
 - add some debugging logs
 - add support to kubernetes version 1.23